# backend/docs/swagger.yaml
basePath: /api
definitions:
  internal_delivery_http.AddPlayerRequest:
    properties:
      device_id:
        type: string
      nickname:
        type: string
      team_id:
        type: string
    type: object
  internal_delivery_http.AddTeamRequest:
    properties:
      color:
        type: string
      name:
        type: string
    type: object
  internal_delivery_http.CreateGameRequest:
    properties:
      settings:
        $ref: '#/definitions/internal_delivery_http.GameSettingsDTO'
    type: object
...
  internal_delivery_http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_delivery_http.EventResponse:
    properties:
      game_id:
        type: string
      id:
        type: string
      payload:
        additionalProperties: {}
        type: object
      timestamp:
        type: string
      type:
        type: string
    type: object
  internal_delivery_http.GameFullResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/internal_delivery_http.EventResponse'
        type: array
      game:
        $ref: '#/definitions/internal_delivery_http.GameResponse'
      players:
        items:
          $ref: '#/definitions/internal_delivery_http.PlayerResponse'
        type: array
      teams:
        items:
          $ref: '#/definitions/internal_delivery_http.TeamResponse'
        type: array
    type: object
  internal_delivery_http.GameResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      ended_at:
        type: string
      id:
        type: string
      settings:
        $ref: '#/definitions/internal_delivery_http.GameSettingsDTO'
      started_at:
        type: string
      status:
        type: string
    type: object
  internal_delivery_http.GameSettingsDTO:
    properties:
      friendly_fire:
        type: boolean
      game_duration:
        type: integer
      max_players:
        type: integer
      respawn_delay:
        type: integer
    type: object
...
  internal_delivery_http.PlayerResponse:
    properties:
      deaths:
        type: integer
      device_id:
        type: string
      game_id:
        type: string
      id:
        type: string
      is_alive:
        type: boolean
      kills:
        type: integer
      nickname:
        type: string
      score:
        type: integer
      team_id:
        type: string
    type: object
  internal_delivery_http.TeamResponse:
    properties:
      color:
        type: string
      game_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  internal_delivery_http.UpdatePlayerTeamRequest:
    properties:
      team_id:
        type: string
    type: object
  internal_delivery_http.UpdateSettingsRequest:
    properties:
      settings:
        $ref: '#/definitions/internal_delivery_http.GameSettingsDTO'
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for the Aimtec 2026 Laser Tag system — manages games, devices,
    players, teams and real-time events.
  title: Laser Tag API
  version: "1.0"
paths:
...
  /games:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http.GameResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all games
      tags:
      - games
    post:
      consumes:
      - application/json
      parameters:
      - description: Game settings (optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.CreateGameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_delivery_http.GameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new game
      tags:
      - games
  /games/{id}:
    get:
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.GameResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get game by ID
      tags:
      - games
  /games/{id}/end:
    post:
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.GameResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: End a game (running → finished)
      tags:
      - games
  /games/{id}/events:
    get:
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http.EventResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get game events (kills, respawns, etc.)
      tags:
      - games
  /games/{id}/full:
    get:
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.GameFullResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get full game state (game + teams + players + events)
      tags:
      - games
  /games/{id}/leaderboard:
    get:
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http.PlayerResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get game leaderboard (players sorted by score)
      tags:
      - games
  /games/{id}/players:
    get:
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http.PlayerResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List players in a game
      tags:
      - players
    post:
      consumes:
      - application/json
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      - description: Player data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.AddPlayerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_delivery_http.PlayerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "409":
          description: Game full or device already in game
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a player (device) to a game
      tags:
      - players
  /games/{id}/players/{playerId}:
    delete:
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      - description: Player ID
        in: path
        name: playerId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a player from a game
      tags:
      - players
  /games/{id}/players/{playerId}/team:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      - description: Player ID
        in: path
        name: playerId
        required: true
        type: string
      - description: New team
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.UpdatePlayerTeamRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change a player's team assignment
      tags:
      - players
  /games/{id}/settings:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      - description: New settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.UpdateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.GameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update game settings (lobby only)
      tags:
      - games
  /games/{id}/start:
    post:
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_delivery_http.GameResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start a game (lobby → running)
      tags:
      - games
  /games/{id}/teams:
    get:
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_delivery_http.TeamResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List teams in a game
      tags:
      - teams
    post:
      consumes:
      - application/json
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      - description: Team data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.AddTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_delivery_http.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a team to a game
      tags:
      - teams
  /games/{id}/teams/{teamId}:
    delete:
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a team
      tags:
      - teams
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" (without quotes).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"